1. Factored common safari displayer code out of bookmarks and articles coordinator into a SafariDisplayer protocol with extension methods 2. One-lined some code 3. Less brittle check for current tab in articles detail view controller
