+++ iOS Client/BookmarksController/BookmarksCoordinator.swift
@@ -11,7 +11,7 @@ import CoreData
 import ReSwift
 import SafariServices
 
-class BookmarksCoordinator: NSObject, Coordinator, StoreSubscriber, SFSafariViewControllerDelegate {
+class BookmarksCoordinator: NSObject, Coordinator, StoreSubscriber, SafariDisplayer {
     var childCoordinators: [Coordinator] = []
     var route: [RouteComponent] = [.base]
     var tab: Tab = .bookmarks
@@ -22,7 +22,7 @@ class BookmarksCoordinator: NSObject, Coordinator, StoreSubscriber, SFSafariView
         return self.navigationController
     }
     
-    private lazy var navigationController: UINavigationController = {
+    lazy var navigationController: UINavigationController = {
         return UINavigationController()
     }()
     
@@ -48,20 +48,6 @@ class BookmarksCoordinator: NSObject, Coordinator, StoreSubscriber, SFSafariView
         build(newRoute: newRoute)
         route = newRoute
         
-        let bookmarksSafariOverlay: URL? = state.navigationState.safariOverlays[.bookmarks]!
-        if self.overlay == nil, let overlay = bookmarksSafariOverlay {
-            self.overlay = overlay
-            let overlayVC = SFSafariViewController(url: overlay)
-            overlayVC.delegate = self
-            self.navigationController.present(overlayVC, animated: true, completion: nil)
-        } else if self.overlay != nil, bookmarksSafariOverlay == nil {
-            self.overlay = nil
-            self.navigationController.dismiss(animated: true, completion: nil)
-        }
-    }
-    
-    // MARK: - SafariViewControllerDelegate
-    func safariViewControllerDidFinish(_ controller: SFSafariViewController) {
-        MBStore.sharedStore.dispatch(SelectedArticleLink(url: nil))
+        self.displaySafariVC(forURL: state.navigationState.safariOverlays[.bookmarks]!)
     }
 }
