+++ iOS Client/Reducers/MBAppReducer.swift
@@ -27,20 +27,12 @@ func navigationReducer(action: Action, state: NavigationState?) -> NavigationSta
     case let action as NavigationActionSwitchTab:
         nextState.selectedTab = action.tab
     case let action as SelectedArticle:
-        if nextState.selectedTab == .articles {
-           nextState.routes[.articles]?.append(.detail(item: action.article))
-        } else if nextState.selectedTab == .bookmarks {
-            nextState.routes[.bookmarks]?.append(.detail(item: action.article))
-        }
+        nextState.routes[nextState.selectedTab]?.append(.detail(item: action.article))
     case let action as SelectedDevotion:
         nextState.selectedTab = .devotions
         nextState.routes[.devotions]?.append(.detail(item: action.devotion))
     case let action as SelectedArticleLink:
-        if nextState.selectedTab == .articles {
-            nextState.safariOverlays[.articles] = action.url
-        } else if nextState.selectedTab == .bookmarks {
-            nextState.safariOverlays[.bookmarks] = action.url
-        }
+        nextState.safariOverlays[nextState.selectedTab] = action.url
     case let action as SelectedPodcast:
         nextState.routes[.podcasts]?.append(.detail(item: action.podcast))
     case _ as PopCurrentNavigation:
